fix(release): replace local dependency protocols of unchanged dependencies #30995 (#33151)
## Current Behavior
When independent versioning with conventional commits, and only one
project in a release group needs bumped, but it depends on another
project via local dependency protocols such as `file://` or `workspace:`
protocol, the local protocol dependency is not replaced.

## Expected Behavior
local protocol is replaced with the current version of the dependency.

## Related Issue(s)

Fixes #30995