fix(core): swc register base url missing when using tsgo
With the new tsgo `baseUrl` has been completely removed, this breaks compilation of plugins using
swc when calling them with nx due to swc needing the baseUrl. I patched my company workspace with
this fix and it did resolve it, not sure if its the right fix or what the ramifications are so happy
to discuss that more. Since `baseUrl` is removed in tsgo (and recommended against in general), we
need to find a way to provide it to swc (potentially through an `.swcrc` alternatively, I tried
adding that to my workspace though and it didn't do anything).