feat(angular): split `vitest` unit test runner choice into `vitest-angular` and `vitest-analog` (#33631)
Splits the `vitest` unit test runner choice in generators into
`vitest-angular` and `vitest-analog`.

The `@angular/build:unit-test` executor requires a build task with
specific executors, so it doesn't work in all setups. This gives users
the choice of what setup they want to use. A migration is provided to
ensure existing generator defaults with `vitest` are switched to
`vitest-analog`, which is the equivalent.

---------

Co-authored-by: nx-cloud[bot] <71083854+nx-cloud[bot]@users.noreply.github.com>