fix(core): exit with error when generator prompts fail (#33691)
## Current Behavior

If there are issues with values passed to generators via prompt, we
still see a green output and exit code 0.

A colleague found out about this by pressing `Ctrl+C` when being
prompted for parameters for a generator, and this led to the CLI simply
continuing execution and showing no issue.

## Expected Behavior

CLI fails.